refactor: modularize component handler
