+++ apps/publish-frt/base/src/builders/templates/mmoomm/playcanvas/handlers/ComponentHandler/components/custom.ts
@@ -1,0 +1,16 @@
+export default function custom(id: string, props: any): string {
+    const safeId = id.replace(/[^a-zA-Z0-9_$]/g, '_');
+    return `
+    // Custom MMO component
+    const customComponent = {
+        properties: ${JSON.stringify(props)},
+
+        applyToEntity(entity) {
+            entity.customComponent_${safeId} = this.properties;
+            console.log('[MMO Component] Custom component ${id} applied to', entity.name);
+        }
+    };
+
+    console.log('[MMO Component] Custom component ${id} ready');
+    `
+}
