+++ apps/publish-frt/base/src/builders/templates/mmoomm/playcanvas/handlers/ComponentHandler/components/physics.ts
@@ -1,0 +1,21 @@
+export default function physics(id: string, props: any): string {
+    return `
+    // Physics component for MMO
+    const physicsComponent = {
+        mass: ${props.mass || 1},
+        friction: ${props.friction || 0.5},
+        restitution: ${props.restitution || 0.3},
+
+        applyToEntity(entity) {
+            entity.addComponent('rigidbody', {
+                type: pc.BODYTYPE_DYNAMIC,
+                mass: this.mass,
+                friction: this.friction,
+                restitution: this.restitution
+            });
+        }
+    };
+
+    console.log('[MMO Component] Physics component ${id} ready');
+    `
+}
